% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_container.R
\docType{data}
\name{data_container}
\alias{data_container}
\title{data_container}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
data_container
}
\description{
data_container
}
\keyword{datasets}
