% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_goalmodel.R
\name{goal_model}
\alias{goal_model}
\title{goal_model}
\usage{
goal_model(goals1, goals2, team1, team2, x1 = NULL, x2 = NULL,
  hfa = TRUE, dc = FALSE, rs = FALSE, fixed_params = NULL,
  weights = NULL, model = "poisson", optim_method = NULL,
  initvals = NULL)
}
\description{
goal_model
}
