% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_models.R
\name{keras_char_cnn_zhang}
\alias{keras_char_cnn_zhang}
\title{keras_char_cnn_zhang}
\usage{
keras_char_cnn_zhang(input_dim, embed_dim = 128, seq_len = 50,
  filter_size = 5, conv_layers = c(100, 60, 30), pool_size = 4,
  dens_layers = c(50, 30, 10), dropout = 0.2, output_dim = 1,
  output_fun = "sigmoid")
}
\arguments{
\item{input_dim}{Number of unique vocabluary/tokens}

\item{embed_dim}{Number of word vectors}

\item{seq_len}{Length of the input sequences}

\item{filter_size}{the window size (kernel_size)}

\item{conv_layers}{number of filters by layer}

\item{pool_size}{pooling dimension (filters)}

\item{dens_layers}{Number of neurons by dense layer}

\item{dropout}{a scalar between 0 and .5}

\item{output_dim}{Number of neurons of the output layer}

\item{output_fun}{Output activation function}
}
\value{
keras model
}
\description{
Character Level Convolutional Neural Network for Text Classification, as described in Zhang et al., 2015 (http://arxiv.org/abs/1509.01626)
https://github.com/mhjabreel/CharCnn_Keras/blob/master/models/char_cnn_zhang.py
}
