% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keras_models.R
\name{keras_sep_cnn}
\alias{keras_sep_cnn}
\title{keras_sep_cnn}
\usage{
keras_sep_cnn(input_dim, embed_dim, seq_len, embed_vectors = NULL,
  n_filters = 30, output_fun = "sigmoid", output_dim = 1)
}
\arguments{
\item{input_dim}{Number of unique vocabluary/tokens}

\item{embed_dim}{Number of word vectors}

\item{seq_len}{Length of the input sequences}

\item{n_filters}{the number of convolutional filters}

\item{output_fun}{Output activation function}

\item{output_dim}{Number of neurons of the output layer}

\item{filter_size}{the window size (kernel_size)}
}
\value{
keras model
}
\description{
keras_sep_cnn
}
