% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\docType{data}
\name{list_metrics}
\alias{list_metrics}
\title{list_metrics[["linear"]]}
\format{An object of class \code{list} of length 4.}
\usage{
list_metrics
}
\description{
list_metrics[["linear"]]
}
\keyword{datasets}
