% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluator.R
\name{rank_prob_score}
\alias{rank_prob_score}
\title{rank_prob_score
https://opisthokonta.net/?p=1333}
\usage{
rank_prob_score(target, probs)
}
\description{
rank_prob_score
https://opisthokonta.net/?p=1333
}
