% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_data.R
\docType{data}
\name{splitter}
\alias{splitter}
\title{#' splitter
#' 
#' @export}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
splitter
}
\description{
#' splitter
#' 
#' @export
}
\keyword{datasets}
