% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_container.R
\docType{data}
\name{text_container}
\alias{text_container}
\title{text_container}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
text_container
}
\description{
text_container
}
\keyword{datasets}
