% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_catboost.R
\docType{data}
\name{trainer_catboost}
\alias{trainer_catboost}
\title{trainer_catboost
 @export}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trainer_catboost
}
\description{
trainer_catboost
 @export
}
\keyword{datasets}
