% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_keras.R
\docType{data}
\name{trainer_keras}
\alias{trainer_keras}
\title{keras_trainer}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trainer_keras
}
\description{
keras_trainer
}
\keyword{datasets}
