% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_ranger.R
\docType{data}
\name{trainer_ranger}
\alias{trainer_ranger}
\title{trainer_ranger
 @export}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trainer_ranger
}
\description{
trainer_ranger
 @export
}
\keyword{datasets}
