% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_rf.R
\docType{data}
\name{trainer_rf}
\alias{trainer_rf}
\title{trainer_rf
 @export}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trainer_rf
}
\description{
trainer_rf
 @export
}
\keyword{datasets}
