% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_rpart.R
\docType{data}
\name{trainer_rpart}
\alias{trainer_rpart}
\title{trainer_rpart
 @export}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trainer_rpart
}
\description{
trainer_rpart
 @export
}
\keyword{datasets}
