% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trainer_xgboost.R
\docType{data}
\name{trainer_xgboost}
\alias{trainer_xgboost}
\title{trainer_xgboost
 @export}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
trainer_xgboost
}
\description{
trainer_xgboost
 @export
}
\keyword{datasets}
